+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Support for reading/writing quantum circuits using OpenQASM
 -   New implementation of the C++ simulator backend based on modern C++ (C++17)
 -   Preliminary support for GPU computations within the C++ simulator backend
+-   Add fSim gate (and related parametric version)
 
 ### Updated